Added stats and fuzzing functionality
Added the option to simultaneously fuzz all cells within a range that
have the same value. This gives better results when formulas contain MIN
and MAX functions. There is a check box which lets you choose which kind
of fuzzing to perform.
Also integrated most of the statistical code we had worked on into this
add-in.