Worked on finding the effect of fuzzing on charts
My thought on this is that charts are not really outputs. They are a
visual representation of all the cells that feed into the chart, and so
the cells themselves are the real outputs. Charts do serve a purpose for
us, because they point out which cells the user cares about (in other
words, which cells are considered output by the user).
Also fixed a minor bug in normalizing the calculated influence.