Correctness and performance improvements; timing
Correctness: Now keeping track of reachability of outputs from cells.
This is used in the standard deviation computation for knowing which
zero-valued entries should count and which should be skipped.
Performance:
1. Reachability table is used in the swapping procedure: Instead of
computing the change in every output, we only look at the changes in
reachable outputs.
2. Jagged array is used for storing z-scores. Every output is only
associated with the cells that are reachable instead of all cells. This
greatly improves the z-score computation performance.
3. The jagged array is also used for looking for outliers. (The large
array is still needed for computing weighted averages of z-scores for
coloring.)
Timing:
Added timers for the different components of the algorithm. Tool outputs
the runtimes at the end.