Fixes in simulation and its output
The simulation didn't work properly with the normal analysis when the
per-worksheet normal analysis flagged a cell that was not an input to
the spreadsheet. There was an assumption that any flagged cells would be
in the inputs, but that isn't the case with the per-worksheet normal
analysis. This was causing an exception because the flagged cell could
not be found in the inputs dictionary.
Also, the simulation output was missing a field in the header row.